requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
dlib==19.15.0
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-SocketIO==3.0.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.3
mysql-connector==2.1.6
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.2.0
pymongo==3.7.1
python-engineio==2.2.0
python-socketio==2.0.0
requests==2.19.1
scipy==1.1.0
shortuuid==0.5.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.2.1
